+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 All notable changes to this project will be documented in this file. This project adheres to
 [Semantic Versioning](http://semver.org/) and [this changelog format](http://keepachangelog.com/).
 
+## Unreleased (2.0)
+
+### Changed
+
+- Added return types to internal methods to remove deprecation messages in PHP 8.1.
+- Minimum `laravel-json-api/core` dependency version is now `2.0`.
+
 ## [1.0.0] - 2021-07-31
 
 Initial stable release, with no changes from `1.0.0-beta.3`.